# Remove `vec_proxy.data.frame()` method