--- a/R/type-data-frame.R
+++ b/R/type-data-frame.R
@@ -81,10 +81,6 @@ vec_ptype_abbr.data.frame <- function(x, ...) {
 }
 
 #' @export
-vec_proxy.data.frame <- function(x, ...) {
-  x
-}
-#' @export
 vec_proxy_compare.data.frame <- function(x, ..., relax = FALSE) {
   x[] <- lapply(x[], vec_proxy_compare, relax = TRUE)
   x