- Removes `vec_proxy.data.frame()`
- Adds `vec_requires_fallback()` which wraps up a common pattern used in slicing/assigning to see if we need to fallback to the R level `[` or `[<-`. In addition to the current checks, it now does _not_ fallback if the type is data frame.

We can see the effects immediately in `vec_size()`:

``` r
library(vctrs)
library(tibble)

df <- data.frame(x = 1, y = 2, z = 3)
tbl <- as_tibble(df)

# before
bench::mark(
  vec_size(df), 
  vec_size(tbl), 
  iterations = 100000
)
#> # A tibble: 2 x 6
#>   expression         min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>    <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_size(df)    1.27µs   1.62µs   526463.    9.83KB     36.9
#> 2 vec_size(tbl)   1.44µs   1.68µs   557824.        0B     27.9

# after
bench::mark(
  vec_size(df), 
  vec_size(tbl), 
  iterations = 100000
)
#> # A tibble: 2 x 6
#>   expression         min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>    <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_size(df)     497ns    636ns  1407780.     6.3KB     42.2
#> 2 vec_size(tbl)    654ns    768ns  1154084.        0B     11.5
```

And more prevalent in `vec_rbind()`:

``` r
library(vctrs)
library(tibble)

df <- data.frame(x = 1, y = 2, z = 3)
tbl <- as_tibble(df)

lst_of_df <- rep_len(list(df), 1e4)
lst_of_tbl <- rep_len(list(tbl), 1e4)

# before
bench::mark(vec_rbind(!!! lst_of_df), iterations = 50)
#> # A tibble: 1 x 6
#>   expression                   min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>              <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_rbind(!!!lst_of_df)   63.4ms   63.4ms      15.8     362KB    1136.

# after
bench::mark(vec_rbind(!!! lst_of_df), iterations = 50)
#> # A tibble: 1 x 6
#>   expression                   min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>              <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_rbind(!!!lst_of_df)   38.9ms   42.2ms      22.5     359KB     59.5


# before
bench::mark(vec_rbind(!!! lst_of_tbl), iterations = 50)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 1 x 6
#>   expression                    min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>               <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_rbind(!!!lst_of_tbl)    224ms    252ms      3.92     377KB     20.8

# after
bench::mark(vec_rbind(!!! lst_of_tbl), iterations = 50)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 1 x 6
#>   expression                    min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>               <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_rbind(!!!lst_of_tbl)    172ms    188ms      5.17     377KB     17.8
```

---

Separate but related notes:

In the process of looking at this, I think I also found an inconsistency in `vec_proxy()`. 

Currently the C level `vec_proxy()` will not ever invoke the `vec_proxy.data.frame()` method for bare data frames, but will for tibbles (as they are considered s3 here).

https://github.com/r-lib/vctrs/blob/390bc5ae59b442928be086c053505d83880025b3/src/proxy.c#L15-L29

On the other hand, the `vec_proxy_info()` function will always invoke the `vec_proxy.data.frame()` method for both bare data frames and tibbles. This is used by many things, like `vec_assert()` (through `vec_is_vector()`) and more importantly, `vec_size()`. That is where it really hurts us, I think.

https://github.com/r-lib/vctrs/blob/390bc5ae59b442928be086c053505d83880025b3/src/type-info.c#L42-L60

I have not done anything to resolve this, but wanted to bring it up