@DavisVaughan 2019-11-14
Decent improvement on detecting the common type too. Mainly improvement in `vec_ptype_common()`. (but we still get hard by dispatch with `vctrs:::vec_ptype2.tbl_df`).

``` r
library(vctrs)
library(tibble)

df <- data.frame(x = 1, y = 2, z = 3)
tbl <- as_tibble(df)

bench::mark(
  vec_ptype2(df, df),
  vec_ptype2(tbl, tbl),
  vec_ptype_common(df, df),
  vec_ptype_common(tbl, tbl),
  check = FALSE,
  iterations = 100000
)

# before

#> # A tibble: 4 x 6
#>   expression                      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                 <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(df, df)           2.61µs    3.4µs   277349.    3.53KB     19.4
#> 2 vec_ptype2(tbl, tbl)         8.64µs   10.9µs    87630.    9.41KB     18.4
#> 3 vec_ptype_common(df, df)     8.16µs    9.1µs   103717.     6.5KB     19.7
#> 4 vec_ptype_common(tbl, tbl)  30.86µs   35.1µs    27407.    7.27KB     26.3

# after

#> # A tibble: 4 x 6
#>   expression                      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                 <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(df, df)           2.73µs   3.51µs   270099.        0B     18.9
#> 2 vec_ptype2(tbl, tbl)         8.62µs  10.52µs    91717.    9.41KB     19.3
#> 3 vec_ptype_common(df, df)     6.25µs   7.05µs   137842.     6.5KB     13.8
#> 4 vec_ptype_common(tbl, tbl)  20.89µs  23.16µs    41860.    7.27KB     15.5
```

<sup>Created on 2019-11-14 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0.9000)</sup>