@DavisVaughan 2019-11-14
Finally, here is the improvement for the tidyr `unnest()` example we've been benchmarking

``` r
library(tidyr)

num_rows <- 10000

tbl <- tibble(
  first = 1:num_rows,
  second = 5:(num_rows+5-1),
  third = 7:(num_rows+7-1)
)

tbl_nest <- tbl %>%
  nest(second_and_third = c(second, third))

df_nest <- as.data.frame(tbl_nest)
df_nest$second_and_third <- lapply(df_nest$second_and_third, as.data.frame)

bench::mark(
  tibble_new = unnest(tbl_nest, second_and_third),
  tibble_old = unnest_legacy(tbl_nest, second_and_third),
  dataframe_new = unnest(df_nest, second_and_third),
  dataframe_old = unnest_legacy(df_nest, second_and_third),
  iterations = 30
)

# before

#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 4 x 6
#>   expression         min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>    <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 tibble_new     660.8ms  740.4ms      1.34    1.43MB     22.1
#> 2 tibble_old      92.2ms   99.1ms     10.1   808.02KB     23.2
#> 3 dataframe_new  329.5ms  407.7ms      2.38  680.05KB     20.6
#> 4 dataframe_old     62ms   67.1ms     14.8   575.67KB     28.7

# after

#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 4 x 6
#>   expression         min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>    <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 tibble_new     467.3ms  569.1ms      1.80    1.43MB     18.7
#> 2 tibble_old      89.7ms   94.4ms     10.5   808.02KB     25.2
#> 3 dataframe_new  262.3ms  340.9ms      2.83  680.05KB     16.7
#> 4 dataframe_old   61.7ms   64.2ms     15.2   575.67KB     18.3
```

<sup>Created on 2019-11-14 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0.9000)</sup>